requirements.txt
streamlit
pandas
numpy
librosa
scikit-learn
matplotlib
sounddevice
SpeechRecognition
gTTS
pygame
Pillow
scipy
